requirements_twilio.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
multidict==6.1.0
pillow==11.0.0
propcache==0.2.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-decouple==3.8
python3-xlib==0.15
pytweening==1.2.0
pywhatkit==5.4
requests==2.32.3
schedule==1.2.2
soupsieve==2.6
twilio==9.4.1
urllib3==2.3.0
Werkzeug==3.1.3
wikipedia==1.4.0
yarl==1.18.3
